deploy/requirements.txt
functions-framework==3.*
flask==2.*
yt-dlp==2023.11.16
requests
python-dotenv
google-cloud-speech
google-cloud-storage
ffmpeg-python